requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
flask==3.0.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing-extensions==4.8.0
werkzeug==3.0.1
wtforms==3.1.1
zipp==3.17.0
sqlalchemyseed==2.0.0
